+++ piece.h
@@ -22,25 +22,25 @@ typedef enum Shape Shape;
 
 struct Piece {
     Shape shape;
-    uint8_t y;
-    uint8_t x;
-    uint8_t n;
-    uint8_t l;
-    uint8_t r;
+    int y;
+    int x;
+    size_t n;
+    size_t l;
+    size_t r;
     bool M[R_MAX][N_MAX][N_MAX];
 };
 
 typedef struct Piece Piece;
 
-Piece piece_get(Shape shape, uint8_t y, uint8_t x);
+Piece piece_get(Shape shape, int y, int x);
 
-Piece* piece_init(Shape shape, uint8_t y, uint8_t x);
+Piece* piece_init(Shape shape, int y, int x);
 
 void piece_destroy(Piece* piece);
 
 void piece_debug_print(Piece* piece);
 
-void piece_move(Piece* piece, uint8_t y, uint8_t x);
+void piece_move(Piece* piece, int y, int x);
 
 void piece_rotate_right(Piece* piece);
